+++ utils/common/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @flatfile/util-common
 
+## 0.2.4
+
+### Patch Changes
+
+- 8a7a0b9: Small upgrade to processRecords() to let the callback return void
+
+## 0.2.3
+
+### Patch Changes
+
+- 28820d5: Release to update @flatfile/api dep
+
 ## 0.2.2
 
 ### Patch Changes