+++ utils/extractor/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @flatfile/plugin-extractor-utils
 
+## 0.4.7
+
+### Patch Changes
+
+- 28820d5: Release to update @flatfile/api dep
+- edaedf5: Fix tick params
+- Updated dependencies [28820d5]
+  - @flatfile/util-file-buffer@0.1.3
+  - @flatfile/util-common@0.2.3
+
 ## 0.4.6
 
 ### Patch Changes